% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_MLE.R
\name{FD_MLE}
\alias{FD_MLE}
\title{Functional Diversity of a single site for specified values of tau and q}
\source{
\url{https://github.com/AnneChao/FunD/blob/master/FunD_Rcode.txt}
}
\usage{
FD_MLE(data, dij, tau, q)
}
\arguments{
\item{data}{a vector of species sample frequencies.}

\item{dij}{a matrix of species-pairwise distances.}

\item{tau}{a numeric for a specified level of threshold distinctiveness.}

\item{q}{a numeric for a specified diversity order q.}
}
\value{
a numeric value of FD.
}
\description{
Calculating functional diversity of a single site for specified values of tau and q based on Chao et al. (2018)
}
\details{
FD_MLE is a function of obtaining FD index of order q. The original R code is ported from \url{https://github.com/AnneChao/FunD/blob/master/FunD_Rcode.txt}.
The function is implemented here for convenience. The details should refer to Chao et al. (2018).
}
\references{
Chao et al. (2018) An attribute-diversity approach to functional diversity, functional beta diversity, and related (dis)similarity measures.
Under revision, Ecological Monographs.
}
