% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunD.R
\name{FunD}
\alias{FunD}
\title{Functional Diversity of N sites for various values of tau and q}
\source{
\url{https://github.com/AnneChao/FunD/blob/master/FunD_Rcode.txt}
}
\usage{
FunD(data, dij, tau, q, boot, datatype)
}
\arguments{
\item{data}{a list with N sites; each element of list is species abundances or species-by-sampling-unit incidence matrix.}

\item{dij}{a matrix of species-pairwise distances.}

\item{tau}{a numeric or a vector of levels of threshold distinctiveness.}

\item{q}{a numeric or a vector of diversity orders; the suggested range for q is [0, 3].}

\item{boot}{a numeric of number of bootstrap replications.}

\item{datatype}{data type of input data: individual-based abundance data (datatype = "abundance") or species-by-sampling-unit incidence matrix (datatype = "incidence_raw").}
}
\value{
two matrices of FD; the first matrix is the q-profile information, the second matrix is the tau profile information.
}
\description{
Calculating functional diversity of N sites for various values of tau and q based on Chao et al. (2018)
}
\details{
FunD is a function of obtaining FD tau-table or q-table. The original R code is ported from \url{https://github.com/AnneChao/FunD/blob/master/FunD_Rcode.txt}.
The function is implemented here for convenience. The details should refer to Chao et al. (2018).
}
\references{
Chao et al. (2018) An attribute-diversity approach to functional diversity, functional beta diversity, and related (dis)similarity measures.
Under revision, Ecological Monographs.
}
