% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema.R
\name{nema}
\alias{nema}
\title{Package for analyzing nematode community structure}
\description{
swtmap package contains simple tools to generate bathymetric map off SW Taiwan.
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>, Jian-Xiang Liao <jianxiangliao@gmail.com >
}
