% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema_abund.R
\docType{data}
\name{nema_abund}
\alias{nema_abund}
\title{Nematode species abundance}
\format{A matrix with cruise, station, deployment, tube, subcore, genus, species and abundance information.}
\description{
Nematode abundance in the Gaopinc Submarine Canyon and the adjacent slope.
}
\details{
Nematode species abundance were estimated from 100 randomly selected individauls from each samples (Liao et al. 2017).
}
\references{
Liao, J.-X., Chen, G.-M., Chiou, M.-D., Jan, S., Wei, C.-L., 2017. Internal tides affect benthic community
structure in an energetic submarine canyon off SW Taiwan. Deep Sea Research Part I: Oceanographic Research Papers 125, 147–160.
}
\keyword{datasets}
