% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema_cruise.R
\docType{data}
\name{nema_cruise}
\alias{nema_cruise}
\title{Environmental data}
\format{A matrix with cruise, station, deployment and environmental data.}
\description{
Environmental data in the Gaopinc Submarine Canyon and the adjacent slope.
}
\details{
Environmental data collected by Liao et al. (2017).
}
\references{
Liao, J.-X., Chen, G.-M., Chiou, M.-D., Jan, S., Wei, C.-L., 2017. Internal tides affect benthic community
structure in an energetic submarine canyon off SW Taiwan. Deep Sea Research Part I: Oceanographic Research Papers 125, 147–160.
}
\keyword{datasets}
