% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema_species.R
\docType{data}
\name{nema_species}
\alias{nema_species}
\title{Nematode functional traits}
\format{A matrix with family, genus, species, buccal morphology, tail shape and life history strategy information.}
\description{
Nematode functional traits in the Gaopinc Submarine Canyon and the adjacent slope.
}
\details{
Nematode functional traits corresponding to specis collected by Liao et al. (2017).
}
\references{
Liao, J.-X., Chen, G.-M., Chiou, M.-D., Jan, S., Wei, C.-L., 2017. Internal tides affect benthic community
structure in an energetic submarine canyon off SW Taiwan. Deep Sea Research Part I: Oceanographic Research Papers 125, 147–160.
}
\keyword{datasets}
