% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema_total.R
\docType{data}
\name{nema_total}
\alias{nema_total}
\title{Total nematode abundance}
\format{A matrix with cruise, station, deployment, tube, subcore and abundance information.}
\description{
Total nematode abundance in the Gaopinc Submarine Canyon and the adjacent slope.
}
\details{
Total nematode abundance in the Gaopinc Submarine Canyon and the adjacent slope was derived from Liao et al. (2017).
}
\references{
Liao, J.-X., Chen, G.-M., Chiou, M.-D., Jan, S., Wei, C.-L., 2017. Internal tides affect benthic community
structure in an energetic submarine canyon off SW Taiwan. Deep Sea Research Part I: Oceanographic Research Papers 125, 147–160.
}
\keyword{datasets}
