% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simper2.R
\name{simper2}
\alias{simper2}
\title{Modified Similarity Percentages (SIMPER) function}
\usage{
simper2(comm, group = rep(1, nrow(comm)))
}
\arguments{
\item{comm}{Community data matrix.}

\item{group}{Factor describing the group structure. Must have at least 2 levels.}
}
\value{
A list of data frame with variables inculding:
\describe{
\item{overall}{The overall within-group dissimilarity}
\item{average}{Average contribution to overall dissimilarity}
\item{sd}{Standard deviation of contribution}
\item{ratio}{Average to sd ratio}
\item{avg}{Average abundances within group}
\item{contr}{\% contribution of speices}
\item{cusum}{Ordered cumulative contribution}
}
}
\description{
Decomposing Bray-Curtis dissimilarity index to estimate the contribution of individual species within group
}
\details{
\link{simper2} modifies from \link[vegan]{simper} function. The original \link[vegan]{simper} function only computes the discriminating species between two groups using Bray-Crutis dissimilarity.
The simper2 decomposes the within-group Bray-Crutis dissimilarity and computes the contribution of each species. One can also achieve the within-group SIMPER with the original \link[vegan]{simper} function by
copying the community data and creasting 2 dummy groups; however, the \link{simper2} should run faster. More details see \link[vegan]{simper}.
}
